+++ tests/unit/framework/web/GroupUrlRuleTest.php
@@ -4,14 +4,14 @@ namespace yiiunit\framework\web;
 
 use yii\web\UrlManager;
 use yii\web\UrlRule;
-use yii\web\PrefixUrlRule;
+use yii\web\GroupUrlRule;
 use yii\web\Request;
 use yiiunit\TestCase;
 
 /**
  * @group web
  */
-class PrefixUrlRuleTest extends TestCase
+class GroupUrlRuleTest extends TestCase
 {
     protected function setUp()
     {
@@ -25,7 +25,7 @@ class PrefixUrlRuleTest extends TestCase
         $suites = $this->getTestsForCreateUrl();
         foreach ($suites as $i => $suite) {
             list ($name, $config, $tests) = $suite;
-            $rule = new PrefixUrlRule($config);
+            $rule = new GroupUrlRule($config);
             foreach ($tests as $j => $test) {
                 list ($route, $params, $expected) = $test;
                 $url = $rule->createUrl($manager, $route, $params);
@@ -41,7 +41,7 @@ class PrefixUrlRuleTest extends TestCase
         $suites = $this->getTestsForParseRequest();
         foreach ($suites as $i => $suite) {
             list ($name, $config, $tests) = $suite;
-            $rule = new PrefixUrlRule($config);
+            $rule = new GroupUrlRule($config);
             foreach ($tests as $j => $test) {
                 $request->pathInfo = $test[0];
                 $route = $test[1];